fix flow error
